refactor!: resolve circular imports
