+++ src/tbp/monty/experiment/sensor_module.py
@@ -11,10 +11,10 @@ from typing import Protocol, Sequence
 
 from typing_extensions import Self
 
-from tbp.monty.frameworks.models.abstract_monty_classes import RuntimeSensorModule
 from tbp.monty.frameworks.sensors import SensorID
 from tbp.monty.sensor_modules import SensorModule
-from tbp.monty.sensor_modules.transforms import Transform
+from tbp.monty.sensor_modules.sensor_module import RuntimeSensorModule
+from tbp.monty.sensor_modules.transforms.transform import Transform
 
 __all__ = [
     "ExperimentSensorModule",
