+++ src/tbp/monty/frameworks/models/abstract_monty_classes.py
@@ -10,10 +10,7 @@
 from __future__ import annotations
 
 import abc
-from typing import Any, Collection, Dict, Protocol, Sequence, TypedDict
-
-import numpy as np
-import numpy.typing as npt
+from typing import Any, Collection, Dict, Protocol, Sequence
 
 from tbp.monty.cmp import Goal, Message
 from tbp.monty.context import RuntimeContext
@@ -30,6 +27,7 @@ from tbp.monty.frameworks.models.motor_system_state import (
 )
 from tbp.monty.frameworks.sensors import SensorID
 from tbp.monty.memento import Memento, Snapshotable
+from tbp.monty.observations import SensorObservation
 
 __all__ = [
     "AgentObservations",
@@ -42,21 +40,7 @@ __all__ = [
     "RuntimeContext",
     "RuntimeLearningModule",
     "SensorModule",
-    "SensorObservation",
 ]
-
-
-class SensorObservation(TypedDict, total=False):
-    """Observations from a sensor."""
-
-    rgba: npt.NDArray[np.uint8]
-    depth: npt.NDArray[np.float64]  # TODO: Verify specific type
-    semantic: npt.NDArray[np.int_]  # TODO: Verify specific type
-    semantic_3d: npt.NDArray[np.int_]  # TODO: Verify specific type
-    sensor_frame_data: npt.NDArray[np.int_]  # TODO: Verify specific type
-    cam_to_world: npt.NDArray[np.float64]  # TODO: Verify specific type
-    pixel_loc: npt.NDArray[np.float64]  # TODO: Verify specific type
-    raw: npt.NDArray[np.uint8]
 
 
 class AgentObservations(Dict[SensorID, SensorObservation]):
@@ -487,45 +471,7 @@ class GoalGenerator(metaclass=abc.ABCMeta):
         pass
 
 
-class RuntimeSensorModule(Protocol):
-    """Monty runtime interface to a Sensor Module."""
-
-    def update_state(self, agent: AgentState) -> None:
-        """Update the proprioceptive state for this Sensor Module.
-
-        Args:
-            agent: The proprioceptive state of this sensor module's Agent.
-        """
-        ...
-
-    def step(
-        self,
-        ctx: RuntimeContext,
-        observation: SensorObservation,
-        motor_only_step: bool = False,
-    ) -> Message | None:
-        """Execute a time-step for the Sensor Module.
-
-        Args:
-            ctx: The runtime context.
-            observation: Sensor observation.
-            motor_only_step: Whether the current step is a motor-only step.
-
-        Returns:
-            An optional percept with features and morphological features.
-        """
-        ...
-
-    def propose_goals(self) -> Collection[Goal]:
-        """Return the goals proposed by this Sensor Module.
-
-        Returns:
-            A collection of proposed Goals.
-        """
-        ...
-
-
-class SensorModule(RuntimeSensorModule, ExperimentSensorModule, metaclass=abc.ABCMeta):
+class SensorModule(ExperimentSensorModule, metaclass=abc.ABCMeta):
     @abc.abstractmethod
     def state_dict(self) -> Memento:
         pass
