+++ src/tbp/monty/frameworks/models/salience/sensor_module.py
@@ -17,7 +17,6 @@ from tbp.monty.cmp import Goal
 from tbp.monty.context import RuntimeContext
 from tbp.monty.frameworks.models.abstract_monty_classes import (
     SensorModule,
-    SensorObservation,
 )
 from tbp.monty.frameworks.models.motor_system_state import AgentState, SensorState
 from tbp.monty.frameworks.models.salience.on_object_observation import (
@@ -31,6 +30,7 @@ from tbp.monty.frameworks.models.salience.strategies import (
 from tbp.monty.frameworks.models.sensor_modules import SnapshotTelemetry
 from tbp.monty.frameworks.sensors import SensorID
 from tbp.monty.memento import Memento
+from tbp.monty.observations import SensorObservation
 
 __all__ = ["SalienceSM"]
