+++ src/tbp/monty/sensor_modules/transforms/gaussian_smoothing.py
@@ -13,7 +13,7 @@ import numpy.typing as npt
 import scipy
 from typing_extensions import Self
 
-from tbp.monty.sensor_modules.transforms import (
+from tbp.monty.sensor_modules.transforms.transform import (
     NoDepthSensorPresent,
     Payload,
     Transform,
