+++ tests/unit/frameworks/models/fakes/sensor_modules.py
@@ -15,10 +15,10 @@ from tbp.monty.cmp import Message
 from tbp.monty.context import RuntimeContext
 from tbp.monty.frameworks.models.abstract_monty_classes import (
     SensorModule,
-    SensorObservation,
 )
 from tbp.monty.frameworks.models.motor_system_state import AgentState
 from tbp.monty.memento import Memento
+from tbp.monty.observations import SensorObservation
 
 __all__ = ["FakeSensorModule"]
