+++ tests/unit/frameworks/models/salience/sensor_module_test.py
@@ -20,7 +20,6 @@ from parameterized import parameterized_class
 
 from tbp.monty.cmp import Goal
 from tbp.monty.context import RuntimeContext
-from tbp.monty.frameworks.models.abstract_monty_classes import SensorObservation
 from tbp.monty.frameworks.models.motor_system_state import AgentState, SensorState
 from tbp.monty.frameworks.models.salience.on_object_observation import (
     OnObjectObservation,
@@ -29,6 +28,7 @@ from tbp.monty.frameworks.models.salience.sensor_module import (
     SalienceSM,
 )
 from tbp.monty.frameworks.sensors import SensorID
+from tbp.monty.observations import SensorObservation
 
 
 class ArrayEqual:
